Move country list logic to CountryList.js in the components folder
